refactor: simplify DefaultItem and pass down rest props
